fix(desktop): guard critical-path installs in useInstallAction
The critical-game-files confirmation lived only in mod-button.tsx's
action callback, so it never fired for the Skins page, which calls
useInstallAction().performInstall directly. Move the confirm dialog
into performInstall itself so both call sites get the same safety
check, and drop the now-redundant block (and unused useConfirm import)
from mod-button.tsx.
